pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "retentioniq"
version = "0.1.0"
description = "Prescriptive retention system for multi-location subscription businesses"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "João Cordova", email = "joao@example.com"},
]

dependencies = [
    # Data
    "pandas>=2.2",
    "polars>=1.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pyarrow>=15.0",
    "dvc>=3.50",

    # ML
    "scikit-learn>=1.5",
    "xgboost>=2.1",
    "lightgbm>=4.4",
    "lifelines>=0.29",
    "shap>=0.45",

    # Causal Inference
    "dowhy>=0.11",
    "econml>=0.15",

    # Optimization
    "pyomo>=6.7",
    "pulp>=2.8",

    # Feature Store
    "feast>=0.39",

    # LLM / Agents
    "langchain>=0.3",
    "langgraph>=0.2",
    "langchain-anthropic>=0.2",
    "langchain-openai>=0.2",
    "pgvector>=0.3",

    # API
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic>=2.9",
    "pydantic-settings>=2.4",

    # Orchestration
    "dagster>=1.8",
    "dagster-webserver>=1.8",

    # Experiment Tracking
    "mlflow>=2.16",

    # Monitoring
    "evidently>=0.4",

    # Data Quality
    "great-expectations>=1.0",

    # Utilities
    "structlog>=24.4",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "jupyterlab>=4.2",
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "ipywidgets>=8.1",
    "nbstripout>=0.7",
    "locust>=2.31",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (no infrastructure required)",
    "integration: Integration tests (requires docker-compose)",
    "slow: Slow tests (model training, optimization)",
]
